+++ packages/components/molecules/f-skeleton-loader/src/components/skeletons/RestaurantCardData.vue
@@ -36,7 +36,7 @@ export default {
 
 .c-restaurantData-text {
     width: 25%;
-    height: spacing(x1.5);
+    height: spacing(c);
 }
 
 </style>