+++ packages/components/organisms/f-header/src/components/Navigation.vue
@@ -698,9 +698,9 @@ export default {
         position: absolute;
         top: 0;
         right: 0;
-        width: spacing(x2) + $nav-featureLinkIcon-width + spacing(x2); // includes padding on both sides
-        height: spacing(x2) + $nav-featureLinkIcon-height + spacing(x2);
-        padding: spacing(x2);
+        width: spacing(d) + $nav-featureLinkIcon-width + spacing(d); // includes padding on both sides
+        height: spacing(d) + $nav-featureLinkIcon-height + spacing(d);
+        padding: spacing(d);
     }
 }