+++ packages/components/pages/f-checkout/src/components/Address.vue
@@ -74,11 +74,11 @@ $address-weight-bold     : $font-weight-bold;
     color: $address-colour;
     @include font-size($address-fontSize);
     font-weight: $address-weight-bold;
-    margin: spacing(x2) 0 spacing();
+    margin: spacing(d) 0 spacing();
 }
 
 .c-address-group {
-    margin: spacing(x2) 0;
+    margin: spacing(d) 0;
     padding: 0;
     border: none;
     @include font-size($address-fontSize);