+++ packages/components/pages/f-loyalty/src/components/NoCardsErrorState.vue
@@ -45,8 +45,8 @@ $stampCard-iconSize: 96px;
 
 .c-noCardsError {
     width: 100%;
-    padding-top: spacing(x0.5);
-    min-height: $stampCard-iconSize + spacing(x0.5); // Icon size plus padding
+    padding-top: spacing(a);
+    min-height: $stampCard-iconSize + spacing(a); // Icon size plus padding
 
     @include media($stampCard-responsive-notMobileViewBreakpoint) {
         margin-left: spacing();