+++ packages/components/pages/f-loyalty/src/components/Terms.vue
@@ -18,7 +18,7 @@
 }
 
 .c-loyalty-terms-link {
-    margin: spacing(x3) 0;
+    margin: spacing(e) 0;
     text-align: center;
     font-weight: $font-body-strong-s-weight;
     text-decoration: none;